+++ Thunderstore/CHANGELOG.md
@@ -1,3 +1,10 @@
+## Version 1.1.0
+
+- Using LethalNetworkAPI for better network communication
+- Improving init and update performance
+- Fix issue where disconnecting players wouldn't drop some items
+- Fix incorrect default value on ResetCooldownOnOrbit
+
 ## Version 1.0.2
 
 - Setting default config to match unmodded game behavior.